==> code/1.5 - Select-Features.ipynb <==
# Select Features Notebook
---

## This notebook is to help select features for future models.

If you want to see benchmarks after selecting, click [here](https://git.generalassemb.ly/laternader/project_2/blob/master/deliverables/code/4%20-%20Modeling-Benchmarks.ipynb) or [here for logged features](https://git.generalassemb.ly/laternader/project_2/blob/master/deliverables/code/4.5%20-%20Log-Benchmarks.ipynb).

##### Skeleton for executing AFTER mapping    
    df = pd.read_csv('./datasets/train_cleaned.csv')
    df_test = pd.read_csv('./datasets/test_cleaned.csv')
    df.drop(columns='Unnamed: 0', inplace=True)
    df_test.drop(columns='Unnamed: 0', inplace=True)
    X_features = ['col1', 'col2']
    X = df[X_features]
    X = pd.get_dummies(data=X, columns=['cat_col1', 'cat_col2'], drop_first=True)
    X_test = df_test[X_features]
    X_test = pd.get_dummies(data=X_test, columns=['cat_col1', 'cat_col2'], drop_first=True)
    y = df['target_col']
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
    lr = LinearRegression()
    lr.fit(X_train, y_train)
    train_pred = lr.predict(X_train)   # This may not be necessary to get a completed model; I think I added it for additional metrics  
    SalePrice = lr.predict(X_test)
    print(SalePrice)

### Import Packages

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

### Import Original Dataframes `train.csv` and `test.csv` 

train = pd.read_csv('datasets/train.csv')
test = pd.read_csv('datasets/test.csv')

test.columns = [name.lower().replace(' ','_') for name in test.columns]
train.columns = [name.lower().replace(' ','_') for name in train.columns]

train.head()

test.head()

# Data dictionary 

<i>This data dictionary is formatted in a way to help me understand what to select for my features. The numbers of nulls come from the original datasets that we were given and do not change. The only thing that changes if it is commented out or not to help me know what features I selected.</i>
    
    <column name> [Number of nulls] (Decision, if needed): small description (number of unique values)

# Data Dictionary
## column_name(Decision)[null values]: description(unique values)

 mssubclass(Add): building class (16)
# mszoning(Ignore)
# lotfrontage[160]: linear feet of street connected to property
# lotarea: lot size in sq ft
# street: type of road access to property (2)
# alley[820]: type of alley access (3)
# lot_shape(add): shape of property (4)
# landcontour(Ignore): flatness of property
# utilities(Add): (4)
# lotconfig(Maybe): (5)
# landslope(Ignore): slope of property(3)
 neighboorhood(Add): street names(25)
# condition1(Maybe): proximity to main(railroad)(9)
# condition2(Maybe): railroad(second)(9)
 bldgtype(Add): building type(5)
 housestyle(Add): style of dwelling(8)
# overallqual(Add): overall material and finish quality(10)
 overallcond(Add): overall condition rating(10)
 yearbuilt(Maybe): 
# yearremodadd: remodel date(same as yearbuilt if none)
# roofstyle: type of roof(6)
# roofmatl: roof material(8)
# exterior1st: exterior covering house(17)
# exterior2nd: second layer(17)
# masvnrtype[1]: masonry veneer type(5)
# masvnrarea[1]: masonry veneer area in sqft
# exterqual: exterior material quality(5)
# foundation: type of foundation(6)
# bsmtqual[25]: height of basement(6)
    ##Ex Excellent (100+ inches)
    ##Gd Good (90-99 inches)
    ##TA Typical (80-89 inches)
    ##Fa Fair (70-79 inches)
    ##Po Poor (<70 inches)
    ##NA No Basement
# bsmtcond[25]: general condition
# bsmtexposure[25]: walkout or garden level basement walls
# bsmtfintype1[25]: quality of basement finished area
# bsmtfinsf1[]: type 1 finished sqft
# bsmtfintype2[25]: second area
# bsmtfinsf2[]: type2 area
# bsmtunfsf[]: unfinished sqft area
# totalbsmtsf[]: total sqft of basement area
#  heating(Add): type of heating(6)
 heatingqc(Add): heating quality and condition(5)
 centralair: central air conditioning(replace yes/no to 1/0)
 electrical(Maybe)[1]: electrical system(5)
# 1stfloorsf: first floor square feet
# 2ndfloorsf: second floor sqft
# lowqualfinsf: low quality finished square feet (all floors)
# grlivarea: above grade (ground) living area sqft
# bsmtfullbath[2]: basement full bathrooms
# bsmthalfbath[2]: basement half bath
fullbath: full bathrooms 
halfbath: half bathrooms
 bedroom(add): number of bedrooms
# kitchen(add): number of kitchens
 kitchenqual(add): kitchen quality(5)
 totrmsabvgrd(add): total rooms above grade (no bathrooms)
 functional(add): home functionality rating(8)
 fireplaces(add)[422]: number of fireplaces
# fireplacequ[1000]: fireplace quality(6)
    ## check if nulls are no fireplace
# garagetype[44]: garage location(7)
# garageyrblt[45]: year garage built
# garagefinish[45]: interior finish of the garage(4)
 garagecars[](add): size of garage in car capacity
#  garagearea[](add): size of garage 
 garagequal[45]: garage quality(6)
 garagecond[45]: garage condition(6)
 paveddrive(add): paved driveway(3)
# wooddecksf: wood deck area in sqft
# openporchsf: open porch area in sqft
# 3ssnporch(ignore): enclosed porch area in sqft
# screenporch(ignore): screen porch area in sqft
# poolarea: pool area in sqft
# poolqc[874]: pool quality(5)
# fence[706]: fence quality(5)
# miscfeature[837]: miscellaneous feature not covered in categories(6)
# miscval: $value of miscellaneous feature 
 mosold(add): month sold
 yrsold(maybe): year sold
 saletype: type of sale(10)

### Feature Selection

The features selected below do not reflect the current features on final presentation. It is purely for organization purposes.

# ctrl + ? easy (un)comment

# Select the features you want to use
features = [
#             'id', 
#             'pid', 
#             'ms_subclass', 
#             'ms_zoning', 
#             'lot_frontage', 
#             'lot_area',
#             'street', 
#             'alley', 
#             'lot_shape', 
#             'land_contour', 
#             'utilities',
#             'lot_config', 
#             'land_slope', 
#             'neighborhood', 
#             'condition_1',
#             'condition_2', 
#             'bldg_type', 
#             'house_style', 
            'overall_qual',
#             'overall_cond', 
            'year_built', 
            'year_remod/add', 
#             'roof_style',
#             'roof_matl', 
#             'exterior_1st',
#             'exterior_2nd',
#             'mas_vnr_type',
            'mas_vnr_area', 
#             'exter_qual', 
#             'exter_cond', 
#             'foundation', 
#             'bsmt_qual',
#             'bsmt_cond', 
#             'bsmt_exposure',
#             'bsmtfin_type_1', 
#             'bsmtfin_sf_1',
#             'bsmtfin_type_2',
#             'bsmtfin_sf_2',
#             'bsmt_unf_sf',
            'total_bsmt_sf',
#             'heating', 
#             'heating_qc', 
#             'central_air', 
#             'electrical', 
            '1st_flr_sf',
#             '2nd_flr_sf',
#             'low_qual_fin_sf', 
            'gr_liv_area',
#             'bsmt_full_bath',
#             'bsmt_half_bath',
            'full_bath',
#             'half_bath',
#             'bedroom_abvgr',
#             'kitchen_abvgr',
#             'kitchen_qual',
#             'totrms_abvgrd',
#             'functional',
#             'fireplaces',
#             'fireplace_qu', 
#             'garage_type',
            'garage_yr_blt',
#             'garage_finish',
            'garage_cars',
            'garage_area',
#             'garage_qual',
#             'garage_cond',
#             'paved_drive',
#             'wood_deck_sf',
#             'open_porch_sf',
#             'enclosed_porch', 
#             '3ssn_porch',
#             'screen_porch',
#             'pool_area',
#             'pool_qc',
#             'fence',
#             'misc_feature',
#             'misc_val',
#             'mo_sold',
#             'yr_sold',
#             'sale_type'
            ]

### Mapping Tracker

# ## WHICH ONES HAVE BEEN MAPPED

#             'id', 
#             'pid', 
            'ms_subclass', 
#             'ms_zoning', 
#             'lot_frontage', 
#             'lot_area',
#             'street', 
#             'alley', 
#             'lot_shape', 
#             'land_contour', 
#             'utilities',
#             'lot_config', 
#             'land_slope', 
            'neighborhood', 
#             'condition_1',
#             'condition_2', 
#             'bldg_type', 
#             'house_style', 
            'overall_qual',
#             'overall_cond', 
#             'year_built', 
#             'year_remod/add', 
#             'roof_style',
#             'roof_matl', 
#             'exterior_1st',
#             'exterior_2nd',
#             'mas_vnr_type',
#             'mas_vnr_area', 
#             'exter_qual', 
#             'exter_cond', 
#             'foundation', 
#             'bsmt_qual',
#             'bsmt_cond', 
#             'bsmt_exposure',
#             'bsmtfin_type_1', 
#             'bsmtfin_sf_1',
#             'bsmtfin_type_2',
#             'bsmtfin_sf_2',
#             'bsmt_unf_sf',
#             'total_bsmt_sf',
            'heating', 
#             'heating_qc', 
#             'central_air', 
            'electrical', 
#             '1st_flr_sf',
#             '2nd_flr_sf',
#             'low_qual_fin_sf', 
#             'gr_liv_area',
#             'bsmt_full_bath',
#             'bsmt_half_bath',
#             'full_bath',
#             'half_bath',
#             'bedroom_abvgr',
#             'kitchen_abvgr',
            'kitchen_qual',
#             'totrms_abvgrd',
            'functional',
#             'fireplaces',
#             'fireplace_qu', 
#             'garage_type',
            'garage_yr_blt',
#             'garage_finish',
#             'garage_cars',
            'garage_area',
            'garage_qual',
#             'garage_cond',
#             'paved_drive',
#             'wood_deck_sf',
#             'open_porch_sf',
#             'enclosed_porch', 
#             '3ssn_porch',
#             'screen_porch',
#             'pool_area',
#             'pool_qc',
#             'fence',
#             'misc_feature',
#             'misc_val',
#             'mo_sold',
#             'yr_sold',
            'sale_type'

#### The code below helped determine categorical columns that needed to be converted into dummies or one-hot encode.

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
# dummies_columns = list(train[X_features].select_dtypes(include='object'))

# Archived Features

These cells contain a history of the features I have used.

## FEATURES 1

features1 = ['ms_subclass',

  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_cond',
  'year_built',

  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']
features1, len(features1)

## FEATURES 2

features2 = ['ms_subclass',
  'street',
  'alley',
  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_qual',
  'overall_cond',
  'year_built',
  'bsmt_qual',
  'bsmt_cond',
  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']
features2, len(features2)

## FEATURES 3

features3 = features
features3, len(features3)

## FEATURES 4

features4 = features
features4, len(features4)

## CHECKING

The information in the next few cells are not indicative of the final work. They are only there to help me check for mistakes or what my selected dataframe looks like.

# Set X = dataframe containing features you want
X = train[features3]
X.head()

X.info()

X.isnull().sum()
