==> code/2 - Test-Cleaning.ipynb <==
# Mapping Notebook for Test Data
---

If you want to see the mapping/cleaning of `train` data, [click here](https://git.generalassemb.ly/laternader/project_2/blob/master/deliverables/code/2%20-%20Train-Cleaning.ipynb).

### Run this cell below if you want to save the cleaned and mapped data to the CSV.

# when done mapping and cleaning
test.to_csv('../data/cleaned_test.csv',index=False)

### Imports

pwd

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

test = pd.read_csv('../data/test.csv')
test.columns = [name.lower().replace(' ','_') for name in test.columns]

# This will rerun to check the changes made through mapping
test.head(5)

# Check the cleaned data
test_cleaned = pd.read_csv('N:\\Documents\\General-Assembly\\dsi-cc13\\projects\\project_2\\deliverables\\data\\cleaned_test.csv')
test_cleaned.sample(5)

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
# dummies_columns = list(test[X_features].select_dtypes(include='object'))

# Mapped ms_subclass
test['ms_subclass']=test['ms_subclass'].map({40:40, # 1 story homes
                         20:40, 
                         30:40, 
                         45:50, # 1-1.5 homes
                         50:50, 
                         60:75, # 2 story
                         70:75, 
                         75:75, 
                         180:190, # family homes
                         190:190,
                         120:160, # PUDs
                         150:160,
                         160:160,
                         90:90, # duplex
                         80:85, # split
                         85:85, 
                        })

# mapping neighborhood
test['neighborhood']=test['neighborhood'].map({'NAmes':'NAmes',
                         'CollgCr':'CollgCr',
                         'OldTown':'OldTown',
                         'Edwards':'Edwards',
                         'Somerst':'Somerst',
                         'NridgHt':'NridgHt',
                         'Gilbert':'Gilbert',
                         'Sawyer':'Sawyer',
                         'NWAmes':'NWAmes',
                         'SawyerW':'SawyerW',
                         'Mitchel':'Mitchel',
                         'BrkSide':'BrkSide',
                         'Crawfor':'Crawfor',
                         'IDOTRR':'IDOTRR',
                         'NoRidge':'NoRidge',
                         'Timber':'Timber',
                         'StoneBr':'StoneBr',
                         'SWISU':'SWISU',
                         'ClearCr':'ClearCr',
                         'MeadowV':'MeadowV',
                         'Blmngtn':'Blmngtn',
                         'BrDale':'BrDale',
                         'NPkVill':'NPkVill',
                         'Veenker':'Veenker',
                         'Blueste':'Blueste',
                         'Greens':'Greens',
                         'GrnHill':'Timber', # not in test set so map to Timber
                         'Landmrk':'CollgCr'}) # next to college via map


# mapping heating qc
test['heating_qc']=test['heating_qc'].map({'Ex':'Ex',
                                             'TA':'TA', 
                                             'Gd':'Gd', 
                                             'Fa':'Fa', 
                                             'Po':'Fa'})

# mapping functional
test['functional']=test['functional'].map({'Typ':'Typ',
                         'Mod':'Mod',
                         'Min2':'Min',
                         'Maj1':'Maj',
                         'Min1':'Min',
                         'Sev':'Maj',
                         'Sal':'Maj',
                         'Maj2':'Maj'})

# mapping electrical
test.electrical=test.electrical.map({np.nan:'Mix',
                    'FuseP':'Fuse',
                     'SBrkr':'Sbrkr',
                     'FuseA':'Fuse',
                     'FuseF':"Fuse"})

# mapping kitchen_qual
test['kitchen_qual']=test['kitchen_qual'].map({'Fa':'Fa',
                          'TA':'TA',
                          'Gd':'Gd',
                          'Ex':'Ex',
                          'Po':'Fa'})

# mapping sale_type
test['sale_type']=test['sale_type'].map({'WD ':'WD',
                       'New':'New',
                       'Con':'Con',
                       'COD':'COD',
                       'VWD':'Oth', # changed to other because not the same as WD
                       'CWD':'CWD',
                       'ConLD':'ConLD',
                       'ConLI':'ConLI',
                       'Oth':'Oth',
                       'ConLw':'ConLw'})

# mapping garage_qual
test['garage_qual']=test['garage_qual'].map({'TA':'TA',
                          'Fa':'Fa',
                          'Gd':'Gd',
                          'Ex':'Gd',
                          'Po':'Po'})

# mapping bsmt_cond

test['bsmt_cond']=test['bsmt_cond'].map({'TA':'Avg',
                        'Gd':'Gd',
                        'Fa':'Avg',
                        'Po':'Avg',
                        'Ex':'Gd'})

# mapping overall_cond
test['overall_qual']=test['overall_qual'].map({6:6,
                           7:7,
                           5:5,
                           8:8,
                           10:10,
                           4:4,
                           9:9,
                           3:3,
                           2:2,
                           1:2})

# map mas_vnr_area, will make 0 because it does not have vnr
test['mas_vnr_area'].fillna(0,inplace=True)

# map total bsmt sf
test['total_bsmt_sf'].fillna(test['total_bsmt_sf'].mean(),inplace=True)

# mapping nulls for garage_yr_blt and garage_area
test['garage_yr_blt'].fillna(np.round(test['garage_yr_blt'].mean()),inplace=True)

test['garage_area']=test['garage_area'].replace(0,np.nan)

test['garage_area'].fillna(np.round(test['garage_area'].mean()),inplace=True)

# mapping 'garage_cars'
test['garage_cars'].fillna(0, inplace=True) # make it 0 because it might not even have a garage
