==> code/3 - Feature-Engineering.ipynb <==
# Feature Engineering 
---

Any features that I thought would be good to create is made in here. All data that comes in here will be after the cleaning. Later at the bottom, that is when I will save the CSVs periodically after creating a feature. 

# Save the CSVs into a featured engineered csv
cleaned_train.to_csv('../data/clean-train-engineered.csv')

cleaned_test.to_csv('../data/clean-test-engineered.csv')

## Import Packages

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

pwd

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Import Cleaned Data

# Import both csv cleaned
cleaned_train =  pd.read_csv('../data/cleaned_train.csv')
cleaned_test = pd.read_csv('../data/cleaned_test.csv')

cleaned_train.head()

cleaned_test.head()

# Check the shape
cleaned_train.shape, cleaned_test.shape

# Let's start some engineering

One obvious feature I wanted to incorporate was the log of the `saleprice` column in the `cleaned_train` dataset.

cleaned_train['log_saleprice'] = cleaned_train['saleprice'].map(np.log)

Since I notice a decent feature between `totrms_abvgrd` with `bedroom_abvgrd` and `full_bath`, I will create an iteraction between them that will be included in the some select feature sets.

# check for any 0s in any of the columns above
cleaned_train[(cleaned_train['totrms_abvgrd'] == 0) | (cleaned_train['bedroom_abvgr']==0) | (cleaned_train['full_bath']==0)]

# Since there are only 4 rows where it applies, then I won't drop any rows.

cleaned_train['totrms * bedroom'] = cleaned_train['totrms_abvgrd'] * cleaned_train['bedroom_abvgr']

# I will add up full_bath and half_bath to create total_bath, not including basement
cleaned_train['total_bath_nobsmt'] = cleaned_train['full_bath'] + cleaned_train['half_bath']
# Then I will multiply it with totrms_abvgrd
cleaned_train['totrms * baths'] = cleaned_train['totrms_abvgrd'] * cleaned_train['total_bath_nobsmt']

# I will also create an interaction of bedrooms, bathrooms, and totrms
cleaned_train['all_rms']=cleaned_train['total_bath_nobsmt'] * cleaned_train['totrms_abvgrd'] * cleaned_train['bedroom_abvgr']

The same will be done to the test set for consistency sake.

cleaned_test['totrms * bedroom'] = cleaned_test['totrms_abvgrd'] * cleaned_test['bedroom_abvgr']
# I will add up full_bath and half_bath to create total_bath, not including basement
cleaned_test['total_bath_nobsmt'] = cleaned_test['full_bath'] + cleaned_test['half_bath']
# Then I will multiply it with totrms_abvgrd
cleaned_test['totrms * baths'] = cleaned_test['totrms_abvgrd'] * cleaned_test['total_bath_nobsmt']
# I will also create an interaction of bedrooms, bathrooms, and totrms
cleaned_test['all_rms']=cleaned_test['total_bath_nobsmt'] * cleaned_test['totrms_abvgrd'] * cleaned_test['bedroom_abvgr']

There wasn't much feature engineering due to it not being needed. The only time it mattered is when introducing the log of the sale price.
