==> code/4.5 - Log-Benchmarks.ipynb <==
# Modeling Benchmarks But with a logged SalePrice

This notebook contains my many models that I run over many different feature sets. We will compare these to the null model as well.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn import metrics

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# Import CSV
clean_train =  pd.read_csv('../data/clean-train-engineered.csv')
clean_test = pd.read_csv('../data/clean-test-engineered.csv')

# Creating a baseline

y = clean_train['saleprice']

baseline = clean_train['saleprice'].mean()
baseline

Since we can't make any predictions right now, we have to go into the [selection notebook](https://git.generalassemb.ly/laternader/project_2/blob/master/deliverables/code/1.5%20-%20Select-Features.ipynb). Since these were logged, the original benchmarks are [here](https://git.generalassemb.ly/laternader/project_2/blob/master/deliverables/code/4%20-%20Modeling-Benchmarks.ipynb).

But for now we can make a null prediction (`baseline` is our null prediction). We need to calculate the distance between a sale price from the 'saleprice' column to the mean of all saleprices. This will give us null residuals.

null_resids = y - baseline
null_resids[:5]

np.mean(null_resids**2)

# This is the null model before the split
clean_train['null_pred'] = baseline
#                                     y before the split  , caluculated mean of y
mse_null = metrics.mean_squared_error(clean_train['saleprice'], clean_train['null_pred'])
mse_null

rmse_null = mse_null**.5
rmse_null

Now we can compare with our models.

---
# Feature Set 1

The features presented here are features that had high positive correlation with 'saleprice' in the train. It will contain LinearRegression, Lasso, And Ridge models

features1 = ['overall_qual',
  'year_built',
  'year_remod/add',
  'mas_vnr_area',
  'total_bsmt_sf',
  '1st_flr_sf',
  'gr_liv_area',
  'full_bath',
  'garage_yr_blt',
  'garage_cars',
  'garage_area']

X_train = clean_train[features1]
X_test = clean_test[features1]
y_train = clean_train['saleprice']

This code helps select any columns that need to be dummified.

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features1].select_dtypes(include='object'))

We will now fit the `train` into its own model before proceeding to apply it to the whole model.

X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69420)

y_trainv_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)

# Instatiate and fit the model
lr = LinearRegression()
lr.fit(X_train, y_trainv_log)

lr.score(X_train, y_trainv_log), lr.score(X_validation, y_valid_log)

preds = lr.predict(X_validation)
preds[:5]

Now that we calculated the R scores, we need to calculate the metrics to compare. In order to do that, we bring in the function below.

# The idea came from lab 3.01
def big_metrics(y, predictions, features):
    # 1. Calculate Sum Squared Error; SSE
    SSE = ((y - predictions)**2).sum()
    
    # 2. Calculate Mean Squared Error; MSE
    MSE = SSE/len(y)
    
    # 3. Calculate Root Mean Squared Error; RMSE
    RMSE = np.sqrt(MSE)
    
    # 4. Calculate Mean Absolute Error; MAE
    MAE = (abs(y - predictions)).sum()/len(y)
    
    # 5. Calculate R2 Score Error; R2
    Null_SSE = ((y - y.mean())**2).sum() # Need null to calculate R2
    R2 = 1 - SSE/Null_SSE
    
    # 6. Calculate R2 Adjusted
    R2_adj = 1-(((1-R2)*(len(y)-1))/(len(y)-len(features)-1))
    
    return SSE, MSE, RMSE, MAE, R2, R2_adj

metric_names = ['SSE', 'MSE', 'RMSE', 'MAE', 'R2', 'R2_adj']

f1_metrics = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(preds), features1),index=metric_names)
f1_metrics

f1_metrics[2], rmse_null

Now time to calculate to compare to the null model.

Since my RMSE of my first model based on `features1` is lower than the rmse_null, that means my model is pimpin.

pd.Series(lr.coef_,index=features1)

pd.Series(np.exp(lr.coef_-1)*100,index=X_train.columns)

There were 2 features that cause a drop in price as their units increase. I think we can do a better model.

# Let's model the first set of features to Ridge and RidgeCV

In order to do that, we need to utilize `StandardScaler`

# Standardize the numbers
X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69420)

y_trainv_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)
# Since we already split it earlier, we will use those to perform on the Ridge model
ss = StandardScaler()
Z_train = ss.fit_transform(X_train)
Z_test = ss.transform(X_validation)

ridge_model = Ridge(alpha=5)
ridge_model.fit(Z_train, y_trainv_log)

print('Ridge Training score:', ridge_model.score(Z_train, y_trainv_log))
print('Ridge Test score:', ridge_model.score(Z_test, y_valid_log)) 

r_alphas = np.logspace(0, 5, 100)
ridge_cv = RidgeCV(alphas=r_alphas, scoring='r2',cv=5)
ridge_cv.fit(Z_train, y_trainv_log);

print('RidgeCV Training score:',ridge_cv.score(Z_train, y_trainv_log))
print('RidgeCV Test score:',ridge_cv.score(Z_test, y_valid_log))

Calculate the RMSE on `Ridge` and `RidgeCV` models.

Z_test_preds = ridge_model.predict(Z_test)
Z_test_preds[:5]

f1_metrics_r = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(Z_test_preds), features1),index=metric_names)
f1_metrics_r

pd.Series(ridge_model.coef_,index=features1)

pd.Series(np.exp(ridge_model.coef_-1)*100,index=X_train.columns)

f1_metrics[2], f1_metrics_r[2], rmse_null

Z_test_preds = ridge_cv.predict(Z_test)
Z_test_preds[:5]

f1_metrics_rcv = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(Z_test_preds), features1),index=metric_names)
f1_metrics_rcv

pd.Series(ridge_cv.coef_,index=features1)

pd.Series(np.exp(ridge_cv.coef_-1)*100,index=X_train.columns)

f1_metrics[2], f1_metrics_r[2], f1_metrics_rcv[2], rmse_null

# We can also LassoCV

l_alphas = np.logspace(1, 0, 100)
# Instantiate
lasso_cv = LassoCV(alphas=l_alphas, cv=5, max_iter=50000)
# Fit
lasso_cv.fit(Z_train, y_trainv_log);

lasso_cv.alpha_

print(lasso_cv.score(Z_train, y_trainv_log))
print(lasso_cv.score(Z_test, y_valid_log))

Z_test_preds = lasso_cv.predict(Z_test)
Z_test_preds[:5]

lasso_cv.coef_

f1_metrics_lasso = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(Z_test_preds), features1),index=metric_names)
f1_metrics_lasso

f1_metrics_lasso[2], f1_metrics[2], f1_metrics_r[2], f1_metrics_rcv[2], rmse_null

pd.Series(lasso_cv.coef_,index=features1)

All of the models did better than the null, but the best one was the LinearRegression model.

Now that we have computed that for our first set of features, the next set of features will follow the same route.

---
# Feature Set 2

# Features
features2 = ['ms_subclass',

  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_cond',
  'year_built',

  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features2].select_dtypes(include='object'))
dummies_columns

X_train = clean_train[features2]
X_test = clean_test[features2]
y_train = clean_train['saleprice']

# Make dummies because we have categorical variables
X_train = pd.get_dummies(data=X_train, columns=dummies_columns)
X_test = pd.get_dummies(data=X_test, columns=dummies_columns)

X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69420)

y_train_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)
# Instatiate and fit the model
lr = LinearRegression()
lr.fit(X_train, y_train_log)

lr.score(X_train,y_train_log), lr.score(X_validation ,y_valid_log)

preds = lr.predict(X_validation)
preds[:5]

f2_metrics = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(preds), features2),index=metric_names)
f2_metrics

f1_metrics[2], f2_metrics[2], rmse_null

pd.Series(lr.coef_,index=X_train.columns)

pd.Series(np.exp(lr.coef_-1)*100,index=X_train.columns)

The second model has gotten worse.

---

# Feature Set 3

features3 = ['ms_subclass',
  'street',
  'alley',
  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_qual',
  'overall_cond',
  'year_built',
  'bsmt_qual',
  'bsmt_cond',
  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features3].select_dtypes(include='object'))
dummies_columns

X_train = clean_train[features3]
X_test = clean_test[features3]
y_train = clean_train['saleprice']

# Make dummies because we have categorical variables
X_train = pd.get_dummies(data=X_train, columns=dummies_columns)
X_test = pd.get_dummies(data=X_test, columns=dummies_columns)

X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69)

y_train_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)
# Instatiate and fit the model
lr = LinearRegression()
lr.fit(X_train, y_train_log)

lr.score(X_train,y_train_log), lr.score(X_validation ,y_valid_log)

preds = lr.predict(X_validation)
preds[:5]

f3_metrics = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(preds), features3),index=metric_names)
f3_metrics

f3_metrics[2], f2_metrics[2], f1_metrics[2], rmse_null 

pd.Series(lr.coef_,index=X_train.columns)

pd.Series(np.exp(lr.coef_-1)*100,index=X_train.columns)

My third model is the better than the second but worse than the first.

---

# Feature Set 4

features4 = ['ms_subclass',
  'street',
  'alley',
  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_qual',
  'overall_cond',
  'year_built',
  'year_remod/add',
  'bsmt_qual',
  'bsmt_cond',
  'total_bsmt_sf',
  'heating_qc',
  'central_air',
  'electrical',
  '1st_flr_sf',
  'gr_liv_area',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_yr_blt',
  'garage_cars',
  'garage_area',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features4].select_dtypes(include='object'))
dummies_columns

X_train = clean_train[features4]
X_test = clean_test[features4]
y_train = clean_train['saleprice']

# Make dummies because we have categorical variables
X_train = pd.get_dummies(data=X_train, columns=dummies_columns)
X_test = pd.get_dummies(data=X_test, columns=dummies_columns)

X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69)

y_train_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)
# Instatiate and fit the model
lr = LinearRegression()
lr.fit(X_train, y_train_log)

lr.score(X_train,y_train_log), lr.score(X_validation ,y_valid_log)

preds = lr.predict(X_validation)
preds[:5]

f4_metrics = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(preds), features4),index=metric_names)
f4_metrics

f4_metrics[2], f3_metrics[2], f2_metrics[2], f1_metrics[2], rmse_null 

pd.Series(lr.coef_,index=X_train.columns)

pd.Series(np.exp(lr.coef_-1)*100,index=X_train.columns)

The numbers continually got worse so I decided to do something different with an old set.

---

# Features Set 2 But some numeric features are treated as categorical

features2s = ['ms_subclass',

  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_cond',
  'year_built',

  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

dummy_columns2s = [
                'neighborhood',
                'bldg_type',
                'house_style',
                'overall_cond',
                'heating_qc',
                'central_air',
                'electrical',
                'kitchen_qual',
                'functional',
                'garage_qual',
                'garage_cond',
                'sale_type']

X_train = clean_train[features2s]
X_test = clean_test[features2s]
y_train = clean_train['saleprice']

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns2s)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns2s)

X_train, X_validation, y_train, y_validation = \
                train_test_split(X_train, y_train, random_state=69)

y_train_log = y_train.map(np.log)
y_valid_log = y_validation.map(np.log)
# Instatiate and fit the model
lr = LinearRegression()
lr.fit(X_train, y_train_log)

lr.score(X_train,y_train_log), lr.score(X_validation ,y_valid_log)

preds = lr.predict(X_validation)
preds[:5]

f2s_metrics = pd.Series(big_metrics(np.exp(y_valid_log), np.exp(preds), features2s),index=metric_names)
f2s_metrics

f2s_metrics[2], f4_metrics[2], f3_metrics[2], f2_metrics[2], f1_metrics[2], rmse_null 

pd.Series(lr.coef_,index=X_train.columns)

pd.Series(np.exp(lr.coef_-1)*100,index=X_train.columns)

It wasn't that far from the original set 2 model results.

I wasn't satisfied with these numbers so I decided to do the same thing but on the logged `saleprice`.

---
