==> code/5 - Submission-Maker.ipynb <==
# SUBMISSION MAKER JUST FEATURES

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn import metrics

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

pwd

# Import CSV
train =  pd.read_csv('../data/clean-train-engineered.csv')
test = pd.read_csv('../data/clean-test-engineered.csv')

# FEATURES1

# Features
features1 = ['ms_subclass',

  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_cond',
  'year_built',

  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']
features1

dummy_columns1 = [
                'neighborhood',
                'bldg_type',
                'house_style',

                'heating_qc',
                'central_air',
                'electrical',
                'kitchen_qual',
                'functional',
                'garage_qual',
                'garage_cond',
                'sale_type']
dummy_columns1

# Create X

# Train X set
X_train = train[features1]

# Test X set
X_test = test[features1]

# Create y_train
y_train = train['saleprice']

# Fill any nulls with 0 
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies / Feature engineering
X_train = pd.get_dummies(data=X_train, columns=dummy_columns1)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns1)

# Set up linear regression
lr1 = LinearRegression()

# fit the model
lr1.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr1.predict(X_test)

y_test[:10]

# now create a column in the test dataframe
test['SalePrice'] = y_test

submissionf1 = test[['id','SalePrice']]
submissionf1.head()

# Rename columns if they are anal about uppercase
submissionf1.rename(columns={'id':'Id'},inplace=True)
submissionf1.columns

# SAVE TO CSV 
submissionf1.to_csv('../submissions/features1.csv', index=False)

# FEATURES 2

features2 = ['ms_subclass',
  'street',
  'alley',
  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_qual',
  'overall_cond',
  'year_built',
  'bsmt_qual',
  'bsmt_cond',
  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

dummy_columns2 = ['street',
                'alley',
                'neighborhood',
                'bldg_type',
                'house_style',
                'bsmt_qual',
                'bsmt_cond',
                'heating_qc',
                'central_air',
                'electrical',
                'kitchen_qual',
                'functional',
                'garage_qual',
                'garage_cond',
                'sale_type']

# Create X

# Train X set
X_train = train[features2]

# Test X set
X_test = test[features2]

# Create y_train
y_train = train['saleprice']

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns2)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns2)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr2 = LinearRegression()

# fit the model
lr2.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr2.predict(X_test)
y_test[:10]

test['SalePrice'] = y_test

submissionf2 = test[['id','SalePrice']]
submissionf2.rename(columns={'id':'Id'},inplace=True)
submissionf2.columns, submissionf2.head()

# SAVE TO CSV 
submissionf2.to_csv('../submissions/features2.csv', index=False)

# FEATURES 3

features3 = ['ms_subclass',
  'street',
  'alley',
  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_qual',
  'overall_cond',
  'year_built',
  'year_remod/add',
  'bsmt_qual',
  'bsmt_cond',
  'total_bsmt_sf',
  'heating_qc',
  'central_air',
  'electrical',
  '1st_flr_sf',
  'gr_liv_area',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_yr_blt',
  'garage_cars',
  'garage_area',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

dummy_columns3 = ['street',
                'alley',
                'neighborhood',
                'bldg_type',
                'house_style',
                'bsmt_qual',
                'bsmt_cond',
                'heating_qc',
                'central_air',
                'electrical',
                'kitchen_qual',
                'functional',
                'garage_qual',
                'garage_cond',
                'sale_type']

# Create X

# Train X set
X_train = train[features3]

# Test X set
X_test = test[features3]

# Create y_train
y_train = train['saleprice']

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns3)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns3)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr3 = LinearRegression()

# fit the model
lr3.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr3.predict(X_test)
y_test[:10]

test['SalePrice'] = y_test

submissionf3 = test[['id','SalePrice']]
submissionf3.rename(columns={'id':'Id'},inplace=True)
submissionf3.columns, submissionf3.head()

# SAVE TO CSV 
submissionf3.to_csv('../submissions/features3.csv', index=False)

# Features 1 but treat some features as categorical

features1 = ['ms_subclass',

  'neighborhood',
  'bldg_type',
  'house_style',
  'overall_cond',
  'year_built',

  'heating_qc',
  'central_air',
  'electrical',
  'full_bath',
  'half_bath',
  'bedroom_abvgr',
  'kitchen_qual',
  'totrms_abvgrd',
  'functional',
  'fireplaces',
  'garage_cars',
  'garage_qual',
  'garage_cond',
  'mo_sold',
  'yr_sold',
  'sale_type']

dummy_columns1 = [
                'neighborhood',
                'bldg_type',
                'house_style',
                'overall_cond',
                'heating_qc',
                'central_air',
                'electrical',
                'kitchen_qual',
                'functional',
                'garage_qual',
                'garage_cond',
                'sale_type']

# Create X

# Train X set
X_train = train[features1]

# Test X set
X_test = test[features1]

# Create y_train
y_train = train['saleprice']

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns1)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns1)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr1d = LinearRegression()
# fit it
lr1d.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr1d.predict(X_test)
y_test[:10]

test['SalePrice'] = y_test


submissionf1d = test[['id','SalePrice']]
submissionf1d.rename(columns={'id':'Id'},inplace=True)
submissionf1d.columns, submissionf1d.head()

# SAVE TO CSV 
submissionf1d.to_csv('../submissions/features1d.csv', index=False)

# Features 4

These features were ones that had a correlation of .5 or higher in relation to saleprice in the `train` set. This was to set a new benchmark and then rise in numbers from here.

features4 = ['overall_qual',
  'year_built',
  'year_remod/add',
  'mas_vnr_area',
  'total_bsmt_sf',
  '1st_flr_sf',
  'gr_liv_area',
  'full_bath',
  'garage_yr_blt',
  'garage_cars',
  'garage_area']

X_train = train[features4]
X_test = test[features4]
y_train = train['saleprice']

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features4].select_dtypes(include='object'))

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummies_columns)
X_test = pd.get_dummies(data=X_test, columns=dummies_columns)

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

#### Instantiate and fit to LinearRegression

lr4 = LinearRegression()
lr4.fit(X_train,y_train)

# now find y_pred or y_test
y_test = lr4.predict(X_test)
y_test[:10]

test['SalePrice'] = y_test


submissionf4 = test[['id','SalePrice']]
submissionf4.rename(columns={'id':'Id'},inplace=True)
submissionf4.columns, submissionf4.head()

# SAVE TO CSV 
submissionf4.to_csv('../submissions/features4.csv', index=False)

# SUBMISSION MAKER FEATURES AND LOG

---

# Features1

# Create X

# Train X set
X_train = train[features1]

# Test X set
X_test = test[features1]

# Create y_train
y_train = train['saleprice'].map(np.log)

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns1)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns1)

# Check which columns are not in the other 
for column in X_train.columns:
    if column not in X_test.columns:
        print(column)
        
print('SPLIT')
for column in X_test.columns:
    if column not in X_train.columns:
        print(column)

# Set up linear regression
lr1log = LinearRegression()

# fit the model
lr1log.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr1log.predict(X_test)

y_test = np.exp(y_test)

# now create a column in the test dataframe
test['SalePrice'] = y_test

submissionf1 = test[['id','SalePrice']]
submissionf1.head()

# Rename columns if they are anal about uppercase
submissionf1.rename(columns={'id':'Id'},inplace=True)
submissionf1.columns

# SAVE TO CSV 
submissionf1.to_csv('../submissions/features1-log.csv', index=False)

# FEATURES 2

# Create X

# Train X set
X_train = train[features2]

# Test X set
X_test = test[features2]

# Create y_train
y_train = train['saleprice'].map(np.log)

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns2)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns2)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr2log = LinearRegression()

# fit the model
lr2log.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr2log.predict(X_test)
y_test = np.exp(y_test)
y_test[:10]

test['SalePrice'] = y_test

submissionf2 = test[['id','SalePrice']]
submissionf2.rename(columns={'id':'Id'},inplace=True)
submissionf2.columns, submissionf2.head()

# SAVE TO CSV 
submissionf2.to_csv('../submissions/features2-log.csv', index=False)

# FEATURES 3

# Create X

# Train X set
X_train = train[features3]

# Test X set
X_test = test[features3]

# Create y_train
y_train = train['saleprice'].map(np.log)

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummy_columns2)
X_test = pd.get_dummies(data=X_test, columns=dummy_columns2)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr3log = LinearRegression()
# fit the model
lr3log.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr3log.predict(X_test)
y_test = np.exp(y_test)
y_test[:10]

test['SalePrice'] = y_test

submissionf3 = test[['id','SalePrice']]
submissionf3.rename(columns={'id':'Id'},inplace=True)
submissionf3.columns, submissionf3.head()

# SAVE TO CSV 
submissionf3.to_csv('../submissions/features3-log.csv', index=False)

# Features 4

These features were ones that had a correlation of .5 or higher in relation to saleprice in the `train` set. This was to set a new benchmark and then rise in numbers from here.

# Create X

# Train X set
X_train = train[features4]

# Test X set
X_test = test[features4]

# Create y_train
y_train = train['saleprice'].map(np.log)

# https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.select_dtypes.html
dummies_columns = list(clean_train[features4].select_dtypes(include='object'))

# Fill any nulls with 0
X_train.fillna(0, inplace=True)
X_test.fillna(0, inplace=True)

# Make dummies
X_train = pd.get_dummies(data=X_train, columns=dummies_columns)
X_test = pd.get_dummies(data=X_test, columns=dummies_columns)

# Check which columns are not in the other 
print('These columns dont exist in test dataframe')
for column in X_train.columns:
    if column not in X_test.columns:
        print('-',column)
        
print('SPLIT')
print('These columns dont exist in train dataframe')
for column in X_test.columns:
    if column not in X_train.columns:
        print('-',column)

# Set up linear regression
lr4log = LinearRegression()
# fit the model
lr4log.fit(X_train, y_train)

# now find y_pred or y_test
y_test = lr4log.predict(X_test)
y_test = np.exp(y_test)
y_test[:10]

test['SalePrice'] = y_test

submissionf4 = test[['id','SalePrice']]
submissionf4.rename(columns={'id':'Id'},inplace=True)
submissionf4.columns, submissionf4.head()

# SAVE TO CSV 
submissionf4.to_csv('../submissions/features4-log.csv', index=False)
